Add support for table grouping and literals
You may now pass a `Channel` as the column, allowing it to specify a column name, an aggregate function (Transform), a single literal value (number or bool), or an array of literals (number or bool).

If aggregates are passed, other columns in the list will be included in a ‘group by’ clause.

If a single literal is passed, we’ll repeat that value for each row in the Data.

If an array of literal is passed, we’ll increment through the literals (row by row), looping through literals if needed.

Literal columns use client side filtering or sorting.

Example:

table(
    penguins,
    columns=[
        "species",
        "flipper_length",
        column(avg("bill_length")),
        column(avg("body_mass")),
        column([1,2,3,4,5,6,7], label=“literal-column)])